+++ mambular/base_models/__init__.py
@@ -1,11 +1,9 @@
-from .classifier import BaseMambularClassifier
-from .distributional import BaseMambularLSS
 from .embedding_classifier import BaseEmbeddingMambularClassifier
 from .embedding_regressor import BaseEmbeddingMambularRegressor
-from .regressor import BaseMambularRegressor
+from .lightning_wrapper import TaskModel
 
-__all__ = ['BaseMambularClassifier',
-           'BaseMambularRegressor',
-           'BaseMambularLSS',
-           'BaseEmbeddingMambularRegressor',
-           'BaseEmbeddingMambularClassifier']
+__all__ = [
+    "BaseEmbeddingMambularRegressor",
+    "BaseEmbeddingMambularClassifier",
+    "TaskModel",
+]
